+++ db/query.sql
@@ -1,12 +1,12 @@
--- name: ListTodos :many
-SELECT * FROM todo ORDER BY id;
+-- name: ListTodosByUser :many
+SELECT id, title FROM todo WHERE user_id = $1 ORDER BY id;
 
 -- name: AddTodo :exec
-INSERT INTO todo (title) VALUES ($1);
+INSERT INTO todo (title, user_id) VALUES ($1, $2);
 
 -- name: RemoveTodo :exec
-DELETE FROM todo WHERE id = $1;
+DELETE FROM todo WHERE id = $1 AND user_id = $2;
 
 -- name: AddTodosBatch :exec
-INSERT INTO todo (title)
-SELECT unnest($1::text[]);
+INSERT INTO todo (title, user_id)
+SELECT unnest(sqlc.arg(titles)::text[]) AS title, sqlc.arg(user_id) AS user_id;
